requirements.txt
einops
opencv-python
tqdm
torchprofile
pycocotools
PyYAML
tensorboard
git+https://github.com/zhijian-liu/torchpack.git@3a5a9f7ac665444e1eb45942ee3f8fc7ffbd84e5
timm==0.9.6
onnx
onnxsim
altair